refactor: extract buildHighlightStyle from mainView into identify/utils
Move highlight style construction out of the monolithic MainView class
into a standalone pure function (issue #1438).
